VTable for overrides
This commit collects all overrides in a single per-type `struct`
emulating a C++ vtable.

Assigning the vtable on construction ensures that only instances created
by the library can have overrides.

The vtable pointer is passed back to the callbacks allowing callers to
allocate additional memory for instance-specific data and a destructor
callback can be used to manage this allocation - in particular, if the
caller needs to allocate closure environments these can be cleaned up
when the instance is destroyed.

A future iteration might separate the vtable into a read-only pointer
struct and a "userdata" pointer for mutable data - alternatively, an
"extrasize" parameter would allocate addtional memory with the instance
itself.