Proof-of-concept nim bindings
This is a trivial POC that reuses the generated C from miqt to create
low-level Nim bindings, enough to run a simple Hello world test.

The high-level bindings in this POC are obviously not ready to be used
(ie there are memory leaks etc) and would need more work (ref vs value
types, inheritance structure, how slots are represented and connected,
etc), probably to bring them closer to nimqt/nimqml/DOtherSide _or_
PySide.

Rough overview of changes needed to the go-based bindings:

* naming fixes - nim and go have different naming conventions - in the
bindings, some of this code currently resides in the C/IL part and
should probably be moved elsewhere
* packaging differences - nim files are imported much like header files
in C meaning it's reasonable to establish a 1:1 relationship between Qt
header and Nim file - this gives a nice "just-works" build experience
where the generated `c++` file is compiled together with the nim project
* types and inheritance structure lives in a separate module to avoid
circular imports (which nim has only limited support for)